+++ c_vector.h
@@ -1,6 +1,8 @@
 #ifndef __C_VECTOR__
 #define __C_VECTOR__
 
+#include <stdint.h>
+
 typedef void *c_vector_h;
 typedef int (*comp_fun)(const void *ele_list, const void *key);
 typedef void (*free_ele)(void *ele);
@@ -26,10 +28,10 @@ struct c_vector_input_init {
 
 c_vector_h c_vector_init(const struct c_vector_input_init *input_args);
 
-int c_vector_free(c_vector_h list);
+uint8_t c_vector_free(c_vector_h list);
 
-int c_vector_insert_in(c_vector_h *list, const void *ele,
-                       const unsigned int index);
+uint8_t c_vector_insert_in(c_vector_h *list, const void *ele,
+                           const unsigned int index);
 
 const void *c_vector_push(c_vector_h *list, const void *ele);
 
@@ -37,16 +39,16 @@ void *c_vector_find(c_vector_h list, const void *ele);
 
 void *c_vector_get_at_index(c_vector_h list, const unsigned int index);
 
-int c_vector_delete_ele(c_vector_h list, const void *ele);
+uint8_t c_vector_delete_ele(c_vector_h list, const void *ele);
 
-int c_vector_delete_ele_at_index(c_vector_h list, const unsigned int index);
+uint8_t c_vector_delete_ele_at_index(c_vector_h list, const unsigned int index);
 
 void c_vector_to_string(c_vector_h list);
 
-unsigned int c_vector_length(const c_vector_h list);
-unsigned int c_vector_capacity(const c_vector_h list);
-unsigned int c_vector_ele_size(const c_vector_h list);
+uint8_t c_vector_length(const c_vector_h list);
+uint8_t c_vector_capacity(const c_vector_h list);
+uint8_t c_vector_ele_size(const c_vector_h list);
 
-unsigned int c_vector_clear(const c_vector_h list);
+uint8_t c_vector_clear(const c_vector_h list);
 
 #endif //!__C_VECTOR__
